chore: update scripts for ipython 8 compatibility
This commit addresses issue #9 by adapting the scripts to the changes introduced in ipython 8, which disables the ipython_config.py trick to prevent potential Execution with Unnecessary Privileges. To mitigate this, the commit introduces the installation of an editable install using `pip install -e .` to make modules inside the source available to the notebook. Additionally, the scripts are updated to use systemctl for restarting, aligning with the changes in the latest SageMaker instances.

Fixes #9